% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitpackage.R
\name{base_freq}
\alias{base_freq}
\title{base frequency}
\usage{
base_freq(fasta)
}
\arguments{
\item{fasta}{}
}
\value{

}
\description{
base frequency
}
\examples{
x<-read_fas("flu_seq.fas")
y<-base_freq(x)
}
