% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitpackage.R
\name{download_GBorFASTA}
\alias{download_GBorFASTA}
\title{Use accn to download the gb files}
\usage{
download_GBorFASTA(acc, database, returntype, returnmode)
}
\arguments{
\item{acc}{accession number}

\item{database}{such as "nucleotide"}

\item{returntype}{gb or fasta}

\item{returnmode}{always choose "text"}
}
\value{
".gb" files
}
\description{
You can choose a accn or many accn to download the gb file
}
\examples{
download_GBorFASTA(paste("AJ5345", 26:49, sep=""), "nucleotide", "gb", "text")
}
