% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitpackage.R
\name{genbank2fasta}
\alias{genbank2fasta}
\title{Convert GB file to FASTA file}
\usage{
genbank2fasta(accn)
}
\arguments{
\item{accn}{}
}
\value{
fasta
}
\description{
Convert GB file to FASTA file
}
\examples{
genbank2fasta("AB115403")
}
