% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitpackage.R
\name{global_align}
\alias{global_align}
\title{Global alignment}
\usage{
global_align(x, y, match, mismatch, gap)
}
\arguments{
\item{x}{seq1}

\item{y}{seq2}

\item{match}{score for match}

\item{mismatch}{score for mismatch}

\item{gap}{score for gap}
}
\value{
result
}
\description{
Global alignment
}
\examples{
global_align(x,y,5,-2,-6)
}
